Split TransitionsPresentersViewModel to TransitionInViewModel and TransitionOutViewModel
