+++ TgaBuilderAvaloniaUi/View/TransitionModeUserControl.axaml
@@ -4,7 +4,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:converter="clr-namespace:TgaBuilderAvaloniaUi.Converters"
              xmlns:ap="clr-namespace:TgaBuilderAvaloniaUi.AttachedProperties"
-             xmlns:transitions="clr-namespace:TgaBuilderLib.Transitions;assembly=TgaBuilderLib"
+             xmlns:tran="clr-namespace:TgaBuilderLib.Transitions;assembly=TgaBuilderLib"
              xmlns:elements="clr-namespace:TgaBuilderAvaloniaUi.Elements"
              mc:Ignorable="d"
              x:CompileBindings="False">
@@ -26,7 +26,7 @@
             Height="36"
             Padding="-16,0,0,0"
             ap:MouseOverInfoAP.InfoText="Transition from top edge"
-            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static transitions:TransitionMode.Top}}"
+            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static tran:TransitionDirection.Top}}"
             Margin="0,0,5,0">
             <Panel Width="20" Height="20">
                 <PathIcon Data="M0,0 L20,0 L20,20 L0,20 Z"
@@ -46,7 +46,7 @@
             Height="36"
             Padding="-16,0,0,0"
             ap:MouseOverInfoAP.InfoText="Transition from right edge"
-            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static transitions:TransitionMode.Right}}"
+            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static tran:TransitionDirection.Right}}"
             Margin="0,0,5,0">
             <Panel Width="20" Height="20">
                 <PathIcon Data="M0,0 L20,0 L20,20 L0,20 Z"
@@ -66,7 +66,7 @@
             Height="36"
             Padding="-16,0,0,0"
             ap:MouseOverInfoAP.InfoText="Transition from bottom edge"
-            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static transitions:TransitionMode.Bottom}}"
+            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static tran:TransitionDirection.Bottom}}"
             Margin="0,0,5,0">
             <Panel Width="20" Height="20">
                 <PathIcon Data="M0,0 L20,0 L20,20 L0,20 Z"
@@ -86,7 +86,7 @@
             Height="36"
             Padding="-16,0,0,0"
             ap:MouseOverInfoAP.InfoText="Transition from left edge"
-            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static transitions:TransitionMode.Left}}"
+            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static tran:TransitionDirection.Left}}"
             Margin="0,0,5,0">
             <Panel Width="20" Height="20">
                 <PathIcon Data="M0,0 L20,0 L20,20 L0,20 Z"
@@ -106,7 +106,7 @@
             Height="36"
             Padding="-16,0,0,0"
             ap:MouseOverInfoAP.InfoText="Transition from diagonal (top-left)"
-            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static transitions:TransitionMode.DiagonalTopLeft}}"
+            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static tran:TransitionDirection.DiagonalTopLeft}}"
             Margin="0,0,5,0">
             <Panel Width="20" Height="20">
                 <PathIcon Data="M0,0 L20,0 L20,20 L0,20 Z"
@@ -126,7 +126,7 @@
             Height="36"
             Padding="-16,0,0,0"
             ap:MouseOverInfoAP.InfoText="Transition from diagonal (top-right)"
-            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static transitions:TransitionMode.DiagonalTopRight}}"
+            IsChecked="{Binding SelectedTransitionMode, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static tran:TransitionDirection.DiagonalTopRight}}"
             Margin="0,0,5,0">
             <Panel Width="20" Height="20">
                 <PathIcon Data="M0,0 L20,0 L20,20 L0,20 Z"
