+++ TgaBuilderAvaloniaUi/View/TransitionWindowControlsUserControl.axaml
@@ -6,7 +6,7 @@
              xmlns:viewmodel="clr-namespace:TgaBuilderLib.ViewModel;assembly=TgaBuilderLib"
              xmlns:ap="clr-namespace:TgaBuilderAvaloniaUi.AttachedProperties"
              xmlns:converters="clr-namespace:TgaBuilderAvaloniaUi.Converters"
-             x:DataType="viewmodel:PivotViewModel"
+             x:DataType="viewmodel:TransitionViewModel"
              x:CompileBindings="False">
     <UserControl.Resources>
         <ResourceDictionary>
@@ -23,7 +23,6 @@
 
             <!-- Mode Selector: Smooth / Brick radio icon buttons -->
             <StackPanel
-                DataContext="{Binding TransitionsPresentersVM}"
                 DockPanel.Dock="Left"
                 Orientation="Horizontal"
                 Margin="0,0,6,0"
@@ -31,11 +30,11 @@
                 <elements:IconRadioButton
                     Icon="{StaticResource tab_in_private}"
                     Classes="outlined"
-                    IsChecked="{Binding SelectedTransitionType, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static viewmodel:TransitionType.Smooth}}"
+                    IsChecked="{Binding IsSmoothMode}"
                     ap:MouseOverInfoAP.InfoText="Smooth transition mode"/>
 
                 <elements:IconRadioButton
-                    IsChecked="{Binding SelectedTransitionType, Converter={StaticResource EnumToBooleanConverter}, ConverterParameter={x:Static viewmodel:TransitionType.Bricks}}"
+                    IsChecked="{Binding IsBrickMode}"
                     Classes="outlined"
                     Margin="5,0,5,0"
                     Icon="{StaticResource table_freeze_regular}"
@@ -44,10 +43,13 @@
             </StackPanel>
 
             <!-- Transition mode buttons panel -->
-            <local:TransitionModeUserControl />
+            <local:TransitionModeUserControl 
+            DataContext="{Binding PivotVM}"
+            />
 
             <!-- Pivot slider -->
             <DockPanel
+                DataContext="{Binding PivotVM}"
                 Margin="5,2,0,0"
                 ap:MouseOverInfoAP.InfoText="Pivot: sets the position of the transition border (0 = left, 1 = right)">
                 <Grid
