+++ TgaBuilderAvaloniaUi/View/TransitionWindowEdgeUserControl.axaml
@@ -19,7 +19,7 @@
 
                         <StackPanel>
                             <StackPanel
-                                DataContext="{Binding TransitionsPresentersVM}"
+                                DataContext="{Binding TransitionInVM}"
                                 Orientation="Horizontal"
                                 Spacing="8">
                                 <elements:IconToggleButton
