+++ TgaBuilderAvaloniaUi/View/TransitionWindowManualUserControl.axaml
@@ -17,7 +17,7 @@
 
     <StackPanel Margin="8,6,8,8">
         <StackPanel
-            DataContext="{Binding TransitionsPresentersVM}"
+            DataContext="{Binding TransitionOutVM}"
             Orientation="Horizontal"
                     Spacing="8"
                     Margin="0,0,0,8">
