+++ TgaBuilderAvaloniaUi/View/TransitionWindowShadowUserControl.axaml
@@ -18,7 +18,7 @@
 
     <StackPanel>
         <StackPanel
-            DataContext="{Binding TransitionsPresentersVM}"
+            DataContext="{Binding TransitionInVM}"
             Orientation="Horizontal"
             Spacing="8">
             <elements:IconToggleButton
