+++ TgaBuilderLib/Transitions/TransitionType.cs
@@ -1,0 +1,7 @@
+namespace Transitions;
+
+public enum TransitionType
+{
+    Smooth,
+    Bricks,
+}
