+++ TgaBuilderLib/ViewModel/Transitions/EdgeViewModel.cs
@@ -7,13 +7,13 @@ public class EdgeViewModel : ThrottledViewModelBase
 {
     public EdgeViewModel(
         ITransitionHelper transitionHelper,
-        TransitionsPresentersViewModel transitionsPresentersVM)
+        TransitionInViewModel transitionInVM)
     {
         _transitionHelper = transitionHelper;
-        TransitionsPresentersVM = transitionsPresentersVM;
+        TransitionInVM = transitionInVM;
     }
 
-    public TransitionsPresentersViewModel TransitionsPresentersVM { get; }
+    public TransitionInViewModel TransitionInVM { get; }
 
     private ITransitionHelper _transitionHelper;
 
@@ -32,6 +32,8 @@ public class EdgeViewModel : ThrottledViewModelBase
         set => SetPropertyTriggerRecalculation(ref _edgeWidth, value);
     }
 
+    public Array EdgeBlendModes => Enum.GetValues(typeof(EdgeBlendMode));
+
     private void ConfigureTransitionHelper()
     {
         _transitionHelper.CurrentBricksPipelineRequirements 
@@ -43,14 +45,14 @@ public class EdgeViewModel : ThrottledViewModelBase
 
     protected override bool PreProcess()
     {
-        return TransitionsPresentersVM.DoPreProcessing();
+        return TransitionInVM.DoPreProcessing();
     }
 
     protected override async Task Recalculate()
     {
         ConfigureTransitionHelper();
  
-        await TransitionsPresentersVM.DoRecalculation();
+        await TransitionInVM.DoRecalculation();
     }
 }
